pyproject.toml
[project]
name = "source-code-knowledge-base"
version = "0.1.0"
description = "Source Code Knowledge Base — RAG knowledge base system for source code architecture analysis, built on LangChain + LangGraph + ChromaDB + Ollama"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "langchain-openai>=0.3",
    "langchain-ollama>=0.3",
    "langgraph>=0.3",
    "chromadb>=1.0",
    "pysqlite3-binary>=0.5",
    "sentence-transformers>=3.0",
    "torch>=2.0",
    "httpx>=0.28",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "typer>=0.15",
    "rich>=13.9",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0",
    "numpy>=2.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.scripts]
sckb = "source_code_kb.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/source_code_kb"]

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.8"]

[tool.uv]
# Use CPU-only PyTorch to avoid downloading ~3.5GB of CUDA libraries
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
